add test module
